bot/requirements.txt
slack-bolt
psycopg2-binary
requests